saas/requirements.txt
# SaaS Podcast Generator Requirements
# Core
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic[email]

# Database
sqlalchemy>=2.0.0

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# Environment
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# Utilities
python-multipart>=0.0.9
email-validator>=2.0.0
